feat: make it possible to render links under the dashboard controls dropdown
