+++ public/app/features/dashboard-scene/scene/DashboardLinksControls.tsx
@@ -1,17 +1,8 @@
-import { sanitizeUrl } from '@grafana/data/internal';
-import { selectors } from '@grafana/e2e-selectors';
 import { sceneGraph } from '@grafana/scenes';
 import { DashboardLink } from '@grafana/schema';
-import { Tooltip } from '@grafana/ui';
-import {
-  DashboardLinkButton,
-  DashboardLinksDashboard,
-} from 'app/features/dashboard/components/SubMenu/DashboardLinksDashboard';
-import { getLinkSrv } from 'app/features/panel/panellinks/link_srv';
-
-import { LINK_ICON_MAP } from '../settings/links/utils';
 
 import { DashboardScene } from './DashboardScene';
+import { DashboardLinkRenderer } from './DashboardLinkRenderer';
 
 export interface Props {
   links: DashboardLink[];
@@ -28,34 +19,11 @@ export function DashboardLinksControls({ links, dashboard }: Props) {
 
   return (
     <>
-      {links.map((link: DashboardLink, index: number) => {
-        const linkInfo = getLinkSrv().getAnchorInfo(link);
-        const key = `${link.title}-$${index}`;
-
-        if (link.type === 'dashboards') {
-          return <DashboardLinksDashboard key={key} link={link} linkInfo={linkInfo} dashboardUID={uid} />;
-        }
-
-        const icon = LINK_ICON_MAP[link.icon];
-
-        const linkElement = (
-          <DashboardLinkButton
-            icon={icon}
-            href={sanitizeUrl(linkInfo.href)}
-            target={link.targetBlank ? '_blank' : undefined}
-            rel="noreferrer"
-            data-testid={selectors.components.DashboardLinks.link}
-          >
-            {linkInfo.title}
-          </DashboardLinkButton>
-        );
-
-        return (
-          <div key={key} data-testid={selectors.components.DashboardLinks.container}>
-            {link.tooltip ? <Tooltip content={linkInfo.tooltip}>{linkElement}</Tooltip> : linkElement}
-          </div>
-        );
-      })}
+      {links
+        .filter((link) => link.placement === undefined || link.placement === 'default')
+        .map((link: DashboardLink, index: number) => (
+          <DashboardLinkRenderer link={link} dashboardUID={uid} key={`${link.title}-$${index}`} />
+        ))}
     </>
   );
 }
